Add rubocop linter (#583)
Adding rubocop to aid in Rails 5 migration and general maintainability.

+++ db/migrate/20120709092821_cleanup.rb
@@ -1,14 +1,12 @@
 class Cleanup < ActiveRecord::Migration
   # Remove obsolete tables and columns that may have been left behind by other migrations.
   def self.up
-    if ActiveRecord::Base.connection.columns('venues').map(&:name).include?("version")
+    if ActiveRecord::Base.connection.columns('venues').map(&:name).include?('version')
       remove_column :venues, :version
     end
 
     %w[event_versions venue_versions].each do |table|
-      if ActiveRecord::Base.connection.tables.include?(table)
-        drop_table table
-      end
+      drop_table table if ActiveRecord::Base.connection.tables.include?(table)
     end
   end
 end
